fix: link addr not updated
